backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
django-cors-headers
django-environ

# dev
djlint
ruff
django-taggit